+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 5.8-SNAPSHOT
 
 ADDED:
  
-- new factory class for create new random objects
+- new factory class for create new random objects with optional exclude fields
 - new factory class for create new SecureRandom objects
 - new methods created for random BigInteger and BigDecimal